+++ lib/data/remote/service/network_logger_service.dart
@@ -1,13 +1,12 @@
 import 'package:application_base/core/service/logger_service.dart';
-import 'package:application_base/core/service/platform_service.dart';
 import 'package:application_base/data/remote/const/request_type.dart';
 import 'package:http/http.dart';
 
-/// Is it possible to send sensitive information to remote logger or not
-bool get _sendSensitive => isDebug;
-
 // TODO(Alex): немного переименовать, как в комментарии
 // https://github.com/AlexSeednov/application_base/pull/3#discussion_r1950461782
+
+/// Can send sensitive data to remote logger or not
+bool canLogSensitive = false;
 
 /// Logging request
 void logRequestSending({
@@ -15,7 +14,7 @@ void logRequestSending({
   required String? body,
 }) {
   String information = 'Request ${request.type} ${request.path} sending';
-  if (_sendSensitive && body != null) information += '\nBody $body';
+  if (canLogSensitive && body != null) information += '\nBody $body';
   logInfo(info: information);
 }
 
@@ -40,7 +39,7 @@ void logResponseGot({
 }) {
   String information = 'Request ${request.type} ${request.path}\n'
       'Response ${response.statusCode}';
-  if (_sendSensitive && response.body.isNotEmpty) {
+  if (canLogSensitive && response.body.isNotEmpty) {
     information += '\nBody ${response.body}';
   }
   logInfo(info: information);
@@ -49,12 +48,11 @@ void logResponseGot({
 /// Error in response
 void logResponseError({
   required RequestType request,
-  required int statusCode,
-  String message = '',
+  required Response response,
 }) {
   String error = 'Request ${request.type} ${request.path}\n'
-      'Response $statusCode';
-  if (message.isNotEmpty) error += ' - $message';
+      'Response ${response.statusCode}';
+  if (response.body.isNotEmpty) error += '\nBody ${response.body}';
   logError(error: error);
 }
 
@@ -66,7 +64,7 @@ void logJsonParsingError({
 }) {
   String error = 'Request $source\n'
       'Got JSON parsing error $info';
-  if (_sendSensitive) error += '\n$json';
+  if (canLogSensitive) error += '\n$json';
   logError(error: error);
 }